Fixed to make sure config override doesnt clobber exiting user config

+++ start.sh
@@ -20,6 +20,9 @@ if [ ! -f /usr/src/workspace/500.html ]
 		cp /tmp/public/500.html /usr/src/workspace/500.html
 fi
 
-cp /usr/src/config/config-override.js /usr/src/workspace/config.js
+if [ ! -f /usr/src/workspace/config.js ]
+	then 
+		cp /usr/src/config/config-override.js /usr/src/workspace/config.js
+fi
 
 npm start
